# Handle response validation error