--- a/lib/reply.js
+++ b/lib/reply.js
@@ -107,7 +107,12 @@ Reply.prototype.send = function (payload) {
     if (hasContentType === false || contentType.indexOf('charset') === -1) {
       this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
     }
-    payload = serialize(this.context, payload, this.res.statusCode)
+    try {
+      payload = serialize(this.context, payload, this.res.statusCode)
+    } catch (error) {
+      onErrorHook(this, error, onSendHook)
+      return
+    }
     flatstr(payload)
   }