--- a/lib/reply.js
+++ b/lib/reply.js
@@ -70,14 +70,14 @@ Reply.prototype.send = function (payload) {
     return
   }
 
-  this.sent = true
-
   if (payload instanceof Error || this[kReplyIsError] === true) {
+    this.sent = true
     onErrorHook(this, payload, onSendHook)
     return
   }
 
   if (payload === undefined) {
+    this.sent = true
     onSendHook(this, payload)
     return
   }
@@ -90,12 +90,14 @@ Reply.prototype.send = function (payload) {
       if (hasContentType === false) {
         this[kReplyHeaders]['content-type'] = CONTENT_TYPE.OCTET
       }
+      this.sent = true
       onSendHook(this, payload)
       return
     }
 
     if (hasContentType === false && typeof payload === 'string') {
       this[kReplyHeaders]['content-type'] = CONTENT_TYPE.PLAIN
+      this.sent = true
       onSendHook(this, payload)
       return
     }
@@ -107,15 +109,11 @@ Reply.prototype.send = function (payload) {
     if (hasContentType === false || contentType.indexOf('charset') === -1) {
       this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
     }
-    try {
-      payload = serialize(this.context, payload, this.res.statusCode)
-    } catch (error) {
-      onErrorHook(this, error, onSendHook)
-      return
-    }
+    payload = serialize(this.context, payload, this.res.statusCode)
     flatstr(payload)
   }
 
+  this.sent = true
   onSendHook(this, payload)
 }