--- a/lib/reply.js
+++ b/lib/reply.js
@@ -71,13 +71,11 @@ Reply.prototype.send = function (payload) {
   }
 
   if (payload instanceof Error || this[kReplyIsError] === true) {
-    this.sent = true
     onErrorHook(this, payload, onSendHook)
     return
   }
 
   if (payload === undefined) {
-    this.sent = true
     onSendHook(this, payload)
     return
   }
@@ -90,14 +88,12 @@ Reply.prototype.send = function (payload) {
       if (hasContentType === false) {
         this[kReplyHeaders]['content-type'] = CONTENT_TYPE.OCTET
       }
-      this.sent = true
       onSendHook(this, payload)
       return
     }
 
     if (hasContentType === false && typeof payload === 'string') {
       this[kReplyHeaders]['content-type'] = CONTENT_TYPE.PLAIN
-      this.sent = true
       onSendHook(this, payload)
       return
     }
@@ -113,7 +109,6 @@ Reply.prototype.send = function (payload) {
     flatstr(payload)
   }
 
-  this.sent = true
   onSendHook(this, payload)
 }
 
@@ -195,6 +190,7 @@ Reply.prototype.callNotFound = function () {
 }
 
 function onSendHook (reply, payload) {
+  reply.sent = true
   if (reply.context.onSend !== null) {
     onSendHookRunner(
       reply.context.onSend,
@@ -305,6 +301,7 @@ function sendStream (payload, res, reply) {
 }
 
 function onErrorHook (reply, error, cb) {
+  reply.sent = true
   if (reply.context.onError !== null && reply[kReplyErrorHandlerCalled] === true) {
     reply[kReplyIsRunningOnErrorHook] = true
     onSendHookRunner(