Fix for issue #1325 : Call onErrorHook if response serialization throw an error


My branch

```
┌─────────┬──────┬──────┬───────┬───────┬─────────┬──────────┬───────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%   │ Avg     │ Stdev    │ Max       │
├─────────┼──────┼──────┼───────┼───────┼─────────┼──────────┼───────────┤
│ Latency │ 0 ms │ 0 ms │ 75 ms │ 81 ms │ 6.82 ms │ 21.98 ms │ 352.12 ms │
└─────────┴──────┴──────┴───────┴───────┴─────────┴──────────┴───────────┘
┌───────────┬─────────┬─────────┬────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%    │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 10111   │ 10111   │ 13383  │ 17567   │ 14242.4 │ 2729.37 │ 10111   │
├───────────┼─────────┼─────────┼────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 1.66 MB │ 1.66 MB │ 2.2 MB │ 2.88 MB │ 2.34 MB │ 448 kB  │ 1.66 MB │
└───────────┴─────────┴─────────┴────────┴─────────┴─────────┴─────────┴─────────┘
```
master
```
┌─────────┬──────┬──────┬───────┬───────┬────────┬──────────┬──────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%   │ Avg    │ Stdev    │ Max      │
├─────────┼──────┼──────┼───────┼───────┼────────┼──────────┼──────────┤
│ Latency │ 0 ms │ 0 ms │ 76 ms │ 82 ms │ 6.8 ms │ 21.86 ms │ 344.8 ms │
└─────────┴──────┴──────┴───────┴───────┴────────┴──────────┴──────────┘
┌───────────┬────────┬────────┬────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 9727   │ 9727   │ 13407  │ 17839   │ 14287.2 │ 3029.42 │ 9725    │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 1.6 MB │ 1.6 MB │ 2.2 MB │ 2.92 MB │ 2.34 MB │ 496 kB  │ 1.59 MB │
└───────────┴────────┴────────┴────────┴─────────┴─────────┴─────────┴─────────┘
```

After fix, performance are 97,2% of actual.

I'm not sure of try-catch solution, but I think it's the less impacting way on the code base. I'm open to better solutions.