@marcobiraghi 2018-12-14
> I think we can move `reply.sent = true` to `onSendHook` and `onErrorHook`.

@mcollina I'm moving `sent = true` in the hooks... but look at here:
https://github.com/fastify/fastify/blob/master/lib/reply.js#L249
https://github.com/fastify/fastify/blob/master/lib/reply.js#L365

It looks like it's not necessary to set `sent = true` in `reply.send()` method, because this is already done later in the flow. I double checked all possible paths but those hooks are always called. So I think the fix can be simply a removal of this line https://github.com/fastify/fastify/blob/master/lib/reply.js#L73