@marcobiraghi 2018-12-15
> Is it passing the tests?

The best solution is to move the `sent = true` inside the hooks, as you said. I will investigate if there is a possibility to improve the `sent` logic, because sometimes it switch from true-false-true many times. But I think this isn't the topic of this PR, so I let the bug fixed only for now.