@mcollina 2018-12-18
I see no difference in performance between this and master.