@marcobiraghi 2018-12-19
@delvedor Do you mean something like this?

```js
fastify.get('/', { schema: { response } }, async function (req, reply) {
  return Promise.resolve()
    .then(() => reply.code(200).send({ work: 'actor' }))
    .catch((error) =>  reply.code(500).send(error))
})
```