@delvedor 2018-12-19
Something like the following should work.
```js
// note that I'm not using an async function
fastify.get('/', { schema: { response } }, function (req, reply) {
  return Promise.resolve({ work: 'actor' })
}
```